CRUD PERMISSION

+++ resources/views/pages/expenses/expenses_index.blade.php
@@ -24,7 +24,9 @@
         <div class="content">
             <div class="content-heading">Expenses
                 <div class="float-right">
-                    <button type="button" onclick="loadFormModal('expense_form', {className: 'Expense'}, 'Create New Expenses', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Expenses</button>
+                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Add Expense"))
+                        <button type="button" onclick="loadFormModal('expense_form', {className: 'Expense'}, 'Create New Expenses', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Expenses</button>
+                    @endif
                 </div>
             </div>
             <div>
@@ -39,7 +41,7 @@
                                     <form  name="expenses_search" action="{{route('expenses_search')}}" id="filter-form" method="post" autocomplete="off">
                                         @csrf
                                         <div class="row">
-                                            <div class="class col-md-3">
+                                            <div class="class col-md-2">
                                                 <div class="input-group mb-3">
                                                     <div class="input-group-prepend">
                                                         <span class="input-group-text" id="basic-addon1">Start</span>
@@ -47,12 +49,25 @@
                                                     <input type="text" name="start_date" id="start_date" class="form-control datepicker-index-form datepicker" aria-describedby="basic-addon1" value="{{date('Y-m-d')}}">
                                                 </div>
                                             </div>
-                                            <div class="class col-md-3">
+                                            <div class="class col-md-2">
                                                 <div class="input-group mb-3">
                                                     <div class="input-group-prepend">
                                                         <span class="input-group-text" id="basic-addon2">End</span>
                                                     </div>
                                                     <input type="text" name="end_date" id="end_date" class="form-control datepicker-index-form datepicker" aria-describedby="basic-addon2" value="{{date('Y-m-d')}}">
+                                                </div>
+                                            </div>
+                                            <div class="class col-md-3">
+                                                <div class="input-group mb-3">
+                                                    <div class="input-group-prepend">
+                                                        <span class="input-group-text" id="basic-addon3">Supervisor</span>
+                                                    </div>
+                                                    <select name="supervisor_id" id="input-supervisor-id" class="form-control" aria-describedby="basic-addon3">
+                                                        <option value="">All</option>
+                                                        @foreach ($supervisors as $supervisor)
+                                                            <option value="{{ $supervisor->id }}"> {{ $supervisor->name }} </option>
+                                                        @endforeach
+                                                    </select>
                                                 </div>
                                             </div>
                                             <div class="class col-md-3">
@@ -84,6 +99,7 @@
                                 <tr>
                                     <th class="text-center" style="width: 100px;">#</th>
                                     <th>Date</th>
+                                    <th>Supervisor Name</th>
                                     <th>Expense Category Name</th>
                                     <th>Description</th>
                                     <th>Amount</th>
@@ -104,6 +120,7 @@
                                             {{$loop->index + 1}}
                                         </td>
                                         <td class="font-w600">{{ $expense->date }}</td>
+                                        <td class="font-w600">{{ $expense->supervisor->name ?? $expense->supervisor_name}}</td>
                                         <td class="font-w600">{{ $expense->expensesCategory->name ?? $expense->category_name }}</td>
                                         <td class="d-none d-sm-table-cell">{{ $expense->description }}
                                         <td class="font-w600">{{ number_format($expense->amount, 2) }}</td>
@@ -116,19 +133,25 @@
                                         </td>
                                         <td class="text-center">
                                             <div class="btn-group">
-                                                <button type="button"
-                                                        onclick="loadFormModal('expense_form', {className: 'Expense', id: {{$expense->id}}}, 'Edit {{$expense->expensesCategory->name ?? $expense->category_name }}', 'modal-md');"
-                                                        class="btn btn-sm btn-primary js-tooltip-enabled"
-                                                        data-toggle="tooltip" title="Edit" data-original-title="Edit">
-                                                    <i class="fa fa-pencil"></i>
-                                                </button>
-                                                <button type="button"
-                                                        onclick="deleteModelItem('Expense', {{$expense->id}}, 'expense-tr-{{$expense->id}}');"
-                                                        class="btn btn-sm btn-danger js-tooltip-enabled"
-                                                        data-toggle="tooltip" title="Delete"
-                                                        data-original-title="Delete">
-                                                    <i class="fa fa-times"></i>
-                                                </button>
+                                                @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Edit Expense"))
+                                                    <button type="button"
+                                                            onclick="loadFormModal('expense_form', {className: 'Expense', id: {{$expense->id}}}, 'Edit {{$expense->supervisor->name ?? $expense->supervisor_name}}', 'modal-md');"
+                                                            class="btn btn-sm btn-primary js-tooltip-enabled"
+                                                            data-toggle="tooltip" title="Edit" data-original-title="Edit">
+                                                        <i class="fa fa-pencil"></i>
+                                                    </button>
+                                                @endif
+
+                                                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Delete Expense"))
+                                                        <button type="button"
+                                                                onclick="deleteModelItem('Expense', {{$expense->id}}, 'expense-tr-{{$expense->id}}');"
+                                                                class="btn btn-sm btn-danger js-tooltip-enabled"
+                                                                data-toggle="tooltip" title="Delete"
+                                                                data-original-title="Delete">
+                                                            <i class="fa fa-times"></i>
+                                                        </button>
+                                                    @endif
+
                                             </div>
                                         </td>
                                     </tr>
