+++ resources/views/pages/gross/gross_index.blade.php
@@ -23,7 +23,10 @@
         <div class="content">
             <div class="content-heading">Gross Profit
                 <div class="float-right">
-                    <button type="button" onclick="loadFormModal('gross_form', {className: 'Gross'}, 'Create New Gross Profit', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Gross Profit</button>
+                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Add Gross Profit"))
+                        <button type="button" onclick="loadFormModal('gross_form', {className: 'Gross'}, 'Create New Gross Profit', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Gross Profit</button>
+                    @endif
+
                 </div>
             </div>
             <div>
@@ -115,19 +118,25 @@
                                         </td>
                                         <td class="text-center">
                                             <div class="btn-group">
-                                                <button type="button"
-                                                        onclick="loadFormModal('gross_form', {className: 'Gross', id: {{$gross->id}}}, 'Edit {{ $gross->supervisor->name ?? $gross->supervisor_name}}', 'modal-md');"
-                                                        class="btn btn-sm btn-primary js-tooltip-enabled"
-                                                        data-toggle="tooltip" title="Edit" data-original-title="Edit">
-                                                    <i class="fa fa-pencil"></i>
-                                                </button>
-                                                <button type="button"
-                                                        onclick="deleteModelItem('Gross', {{$gross->id}}, 'gross-tr-{{$gross->id}}');"
-                                                        class="btn btn-sm btn-danger js-tooltip-enabled"
-                                                        data-toggle="tooltip" title="Delete"
-                                                        data-original-title="Delete">
-                                                    <i class="fa fa-times"></i>
-                                                </button>
+                                                @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Edit Gross Profit"))
+                                                    <button type="button"
+                                                            onclick="loadFormModal('gross_form', {className: 'Gross', id: {{$gross->id}}}, 'Edit {{ $gross->supervisor->name ?? $gross->supervisor_name}}', 'modal-md');"
+                                                            class="btn btn-sm btn-primary js-tooltip-enabled"
+                                                            data-toggle="tooltip" title="Edit" data-original-title="Edit">
+                                                        <i class="fa fa-pencil"></i>
+                                                    </button>
+                                                @endif
+
+                                                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Delete Gross Profit"))
+                                                        <button type="button"
+                                                                onclick="deleteModelItem('Gross', {{$gross->id}}, 'gross-tr-{{$gross->id}}');"
+                                                                class="btn btn-sm btn-danger js-tooltip-enabled"
+                                                                data-toggle="tooltip" title="Delete"
+                                                                data-original-title="Delete">
+                                                            <i class="fa fa-times"></i>
+                                                        </button>
+                                                    @endif
+
                                             </div>
                                         </td>
                                     </tr>
