+++ resources/views/pages/settings/settings_allowance_subscriptions.blade.php
@@ -23,7 +23,10 @@
         <div class="content">
             <div class="content-heading">Settings
                 <div class="float-right">
-                    <button type="button" onclick="loadFormModal('settings_allowance_subscriptions_form', {className: 'AllowanceSubscription'}, 'Create New AllowanceSubscription', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New AllowanceSubscription</button>
+                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Add Allowance Subscription"))
+                        <button type="button" onclick="loadFormModal('settings_allowance_subscriptions_form', {className: 'AllowanceSubscription'}, 'Create New AllowanceSubscription', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10">
+                            <i class="si si-plus">&nbsp;</i>New AllowanceSubscription</button>@endif
+
                 </div>
             </div>
             <div>
@@ -63,12 +66,17 @@
                                     </td>
                                     <td class="text-center" >
                                         <div class="btn-group">
-                                            <button type="button" onclick="loadFormModal('settings_allowance_subscriptions_form', {className: 'AllowanceSubscription', id: {{$allowance_subscription->id}}}, 'Edit {{$allowance_subscription->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
-                                                <i class="fa fa-pencil"></i>
-                                            </button>
-                                            <button type="button" onclick="deleteModelItem('AllowanceSubscription', {{$allowance_subscription->id}}, 'allowance_subscription-tr-{{$allowance_subscription->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
-                                                <i class="fa fa-times"></i>
-                                            </button>
+                                            @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Edit Allowance Subscription"))
+                                                <button type="button" onclick="loadFormModal('settings_allowance_subscriptions_form', {className: 'AllowanceSubscription', id: {{$allowance_subscription->id}}}, 'Edit {{$allowance_subscription->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
+                                                    <i class="fa fa-pencil"></i>
+                                                </button>
+                                            @endif
+                                                @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Delete Allowance Subscription"))
+                                                    <button type="button" onclick="deleteModelItem('AllowanceSubscription', {{$allowance_subscription->id}}, 'allowance_subscription-tr-{{$allowance_subscription->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
+                                                        <i class="fa fa-times"></i>
+                                                    </button>
+                                                @endif
+
                                         </div>
                                     </td>
                                 </tr>
