+++ resources/views/pages/settings/settings_deductions.blade.php
@@ -23,7 +23,10 @@
         <div class="content">
             <div class="content-heading">Settings
                 <div class="float-right">
-                    <button type="button" onclick="loadFormModal('settings_deduction_form', {className: 'Deduction'}, 'Create New Deduction', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Deduction</button>
+                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Add Deduction"))
+                        <button type="button" onclick="loadFormModal('settings_deduction_form', {className: 'Deduction'}, 'Create New Deduction', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10">
+                            <i class="si si-plus">&nbsp;</i>New Deduction</button>@endif
+
                 </div>
             </div>
             <div>
@@ -59,12 +62,17 @@
                                         <td class="font-w600">{{ $deduction->registration_number }}</td>
                                         <td class="text-center" >
                                             <div class="btn-group">
-                                                <button type="button" onclick="loadFormModal('settings_deduction_form', {className: 'Deduction', id: {{$deduction->id}}}, 'Edit {{$deduction->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
-                                                    <i class="fa fa-pencil"></i>
-                                                </button>
-                                                <button type="button" onclick="deleteModelItem('deduction', {{$deduction->id}}, 'deduction-tr-{{$deduction->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
-                                                    <i class="fa fa-times"></i>
-                                                </button>
+                                                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Edit Deduction"))
+                                                    <button type="button" onclick="loadFormModal('settings_deduction_form', {className: 'Deduction', id: {{$deduction->id}}}, 'Edit {{$deduction->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
+                                                        <i class="fa fa-pencil"></i>
+                                                    </button>
+                                                @endif
+                                                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Delete Deduction"))
+                                                        <button type="button" onclick="deleteModelItem('deduction', {{$deduction->id}}, 'deduction-tr-{{$deduction->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
+                                                            <i class="fa fa-times"></i>
+                                                        </button>
+                                                    @endif
+
                                             </div>
                                         </td>
                                     </tr>
