+++ resources/views/pages/settings/settings_staff_salaries.blade.php
@@ -23,7 +23,9 @@
         <div class="content">
             <div class="content-heading">Settings
                 <div class="float-right">
-                    <button type="button" onclick="loadFormModal('settings_staff_salary_form', {className: 'StaffSalary'}, 'Create New StaffSalary', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New StaffSalary</button>
+                    @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Add Staff Salary"))
+                        <button type="button" onclick="loadFormModal('settings_staff_salary_form', {className: 'StaffSalary'}, 'Create New StaffSalary', 'modal-md');" class="btn btn-rounded btn-outline-primary min-width-125 mb-10"><i class="si si-plus">&nbsp;</i>New Staff Salary</button>
+                    @endif
                 </div>
             </div>
             <div>
@@ -59,12 +61,18 @@
                                     </td>
                                     <td class="text-center" >
                                         <div class="btn-group">
-                                            <button type="button" onclick="loadFormModal('settings_staff_salary_form', {className: 'StaffSalary', id: {{$staff_salary->id}}}, 'Edit {{$staff_salary->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
-                                                <i class="fa fa-pencil"></i>
-                                            </button>
-                                            <button type="button" onclick="deleteModelItem('StaffSalary', {{$staff_salary->id}}, 'staff_salary-tr-{{$staff_salary->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
-                                                <i class="fa fa-times"></i>
-                                            </button>
+                                            @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Edit Staff Salary"))
+                                                <button type="button" onclick="loadFormModal('settings_staff_salary_form', {className: 'StaffSalary', id: {{$staff_salary->id}}}, 'Edit {{$staff_salary->name}}', 'modal-md');" class="btn btn-sm btn-primary js-tooltip-enabled" data-toggle="tooltip" title="Edit" data-original-title="Edit">
+                                                    <i class="fa fa-pencil"></i>
+                                                </button>
+                                            @endif
+
+                                                @if(\App\Models\UsersPermission::isUserAllowed(Auth::user()->id,"CRUD","Delete Staff Salary"))
+                                                    <button type="button" onclick="deleteModelItem('StaffSalary', {{$staff_salary->id}}, 'staff_salary-tr-{{$staff_salary->id}}');" class="btn btn-sm btn-danger js-tooltip-enabled" data-toggle="tooltip" title="Delete" data-original-title="Delete">
+                                                        <i class="fa fa-times"></i>
+                                                    </button>
+                                                @endif
+
                                         </div>
                                     </td>
                                 </tr>
